Bill S6375 version original - Kevin S. Parker
BILL NUMBER: S6375 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing the division of minority mental health within the office of mental health, and providing for its powers and duties PURPOSE: The purpose of this bill is to evaluate and improve the current mental health service delivery model for ethnic and minority populations in New York State. SUMMARY OF PROVISIONS: Section 1: Legislative intent. Section 2: Short title.