Bill S6388 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S6388 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to delaying the start date of the operation and enforcement of the central business district tolling program PURPOSE: This bill will delay the implementation of the Central Business District Tolling Program, also referred to as "Congestion Pricing" until January 1, 2029. SUMMARY OF PROVISIONS: Section 1 amends subdivision 8 of section 1703 of the vehicle and traf- fic law.