Bill S6412 version original - Timothy M. Kennedy
BILL NUMBER: S6412 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to requiring trucks to utilize commercial GPS PURPOSE OR GENERAL IDEA OF BILL: To mandate the use of commercial gps in trucks on public highways in New York State that accounts for height, width, length, and weight. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds section 379 to the Vehicle and Traffic Law, providing definitions and penalties for violations.