Bill S6396 version original - Kevin S. Parker
BILL NUMBER: S6396 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to instruction in food, agriculture and nutrition PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to include instruction on food, agri- culture, and nutrition in the statewide health curriculum and to promote healthy eating behavior in students. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The education law is amended by adding a new section 804-e.