Bill S6397 version original - Kevin S. Parker
BILL NUMBER: S6397 SPONSOR: PARKER TITLE OF BILL: An act to amend the election law, in relation to qualifications of voters convicted of a felony; and repealing certain provisions of such law relating thereto PURPOSE OR GENERAL IDEA OF BILL: To provide re-enfranchisement for voters with felony convictions. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivisions 2,3,4,and 5 of section 5-106 of the election law are REPEALED and subdivision 6 is renumbered subdivision 2.