Bill S6399 version original - Kevin S. Parker
BILL NUMBER: S6399 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to enacting the "NY Cares Act" PURPOSE OR GENERAL IDEA OF BILL: Maximize the ability of student service members and their family members to transition to college. Create models for consistently, transparently, and effectively awarding credit for military training and experience that can be scaled regionally and nationally.