Bill S6384 version original - Kevin S. Parker
BILL NUMBER: S6384 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to establishing certain duties for pharmacies when pharmacists employed by the pharmacies refuse to fill prescriptions on the basis of personal beliefs PURPOSE: To insure that an individual's right to religious beliefs and worship cannot impede an individual's legal access to legal prescriptions, including contraception. SUMMARY OF PROVISIONS: Section 1: Legislative findings.