Bill S6405 version original - Mark Walczyk
BILL NUMBER: S6405 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the public officers law, in relation to waiving the residency requirement for the Herkimer county commissioner of social services PURPOSE OR GENERAL IDEA OF BILL: To waive the residency requirement for the Commissioner of Social Services for Herkimer County. SUMMARY OF PROVISIONS: Section 1 - Section 3 of the public officers law is amended by adding a new subdivision 73 to read as follows.