Bill S6369 version original - Kevin S. Parker
BILL NUMBER: S6369 SPONSOR: PARKER TITLE OF BILL: An act to amend the public housing law, in relation to low income hous- ing tax credit eligibility PURPOSE: The bill expands the eligibility threshold for tenants in state LIHTC multi-family developed properties to up to 125% of AMI or less. JUSTIFICATION: The AMI for the City of New York as of December 2013 increased to $63,000 per annum. Currently forty-six percent of New Yorkers are considered poor or almost poor.