Bill S4413A version A - Jessica Ramos
BILL NUMBER: S4413A SPONSOR: RAMOS TITLE OF BILL: An act to amend the tax law, in relation to excise taxes on premium cigars; and providing for the repeal of such provisions upon expiration thereof PURPOSE: To establish the rate of the tobacco products tax exclusively for premi- um cigars at seventy-five percent of the wholesale price or fifty cents, whichever is less. SUMMARY OF PROVISIONS: Section 1 establishes the definition of "premium cigar".