Bill S6389 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S6389 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the general municipal law, in relation to requiring a public hearing to be held when land is proposed to be converted for public use PURPOSE: To allow community members to voice concerns and provide input when private property in their community is proposed for conversion to public use. SUMMARY OF PROVISIONS: Section 1 amends the general municipal law by adding a new section 96-aa. Section 2 provides an effective date.