Bill S6394 version original - Kevin S. Parker
BILL NUMBER: S6394 SPONSOR: PARKER TITLE OF BILL: An act to amend the criminal procedure law, in relation to admissibility of a victim's sexual conduct in a sex offense PURPOSE: To include sex workers under the rape shield law when they become victims of a sexual assault. SUMMARY OF PROVISIONS: Section 1: Section 60.42 of the criminal procedure law, as amended by section 1 of part R of chapter 55 of the laws of 2019, is amended.