Bill S6378 version original - Kevin S. Parker
BILL NUMBER: S6378 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to job creation tax credits PURPOSE OR GENERAL IDEA OF BILL: This bill provides a tax credit for any business which creates new employment in New York State. The credit will be 500 of the new income tax revenue created as a result of the additional jobs created. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Short title. Section 2: Section 210-B of the tax law is amended by adding a new subdivision 59.