Bill S6366 version original - Kevin S. Parker
BILL NUMBER: S6366 SPONSOR: PARKER TITLE OF BILL: An act to amend the penal law, in relation to bias-related graffiti SUMMARY OF PROVISIONS: Section 1: Subdivision 3 of section 485.05 of the penal law, as amended by section 3 of part R of chapter 55 of the laws of 2020, is amended. Section 2: This act shall take effect on the first of November next succeeding the date on which it shall have become a law.