Bill S6382 version original - Kevin S. Parker
BILL NUMBER: S6382 SPONSOR: PARKER TITLE OF BILL: An act to amend the election law, in relation to providing voting mate- rials and assistance in Haitian Creole PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to allow for any Haitian Creole speakers to have an easier understanding of the election process. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The election law is amended by adding a new section 3-507.