Bill S5986A version A - Iwen Chu
BILL NUMBER: S5986A SPONSOR: CHU TITLE OF BILL: An act to amend the education law, in relation to digital materials PURPOSE: The purpose of this bill is to amend the education law to include digital materials when in reference to school library materials. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Amends subdivision 2 of § 711-A(1) to define school library materials as all audio/visual and printed materials, and now includes those in digital format.