Bill S6437 version original - Brad Hoylman-Sigal
BILL NUMBER: S6437 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the general obligations law, in relation to the liabil- ity of electronic marketplaces for compensation for damages for defec- tive goods SUMMARY OF PROVISIONS: Section 1 adds a new section 11-108 to the general obligations law in which: 1) Defines the following terms: a. Electronic marketplace provider b. Handmade c. Marketplace seller d.