Bill S3112A version A - John Mannion
BILL NUMBER: S3112A SPONSOR: MANNION TITLE OF BILL: An act to amend the public buildings law and the public health law, in relation to requiring public institutions and buildings to be equipped with opioid antagonists PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to require that public institutions and buildings are equipped with opioid antagonists. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill would establish that this act shall be known as "Erin's Law".