Bill S6418 version original - Jack M. Martins
BILL NUMBER: S6418 SPONSOR: MARTINS TITLE OF BILL: An act to amend the general business law and the executive law, in relation to the regulation of social media companies and social media platforms PURPOSE: This bill is intended to make it unlawful for a minor to hold or open a social media account without parental consent and access. SUMMARY OF PROVISIONS: Section 1. amends the general business law by adding a new article 9-D; referred to as the New York Social Media Regulation Act.