Bill S6419E version original - Brian Kavanagh
BILL NUMBER: S6419 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to establishing a mattress collection program PURPOSE OF BILL: This bill would require producer responsibility for the collection and recycling of mattresses in New York State. SUMMARY OF PROVISIONS: Section one of this bill amends the environmental conservation law to add a new title 34 that would create a statewide mattress collection program.