Bill S6443 version original - Michelle Hinchey
BILL NUMBER: S6443 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to changing the duration of certain licenses PURPOSE: This bill would change the duration of brewers and farm brewers licenses from one year to time years. SUMMARY OF PROVISIONS: Section 1 amends section 57-a of the alcoholic beverage control law to authorize the State Liquor Authority to extend the duration of brewers and farm brewers licenses to a maximum of three years.