Bill S6458 version original - Monica Martinez
BILL NUMBER: S6458 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the social services law, in relation to certain calls received by child protective services PURPOSE: The purpose of this bill is to provide additional attention to reports of child abuse or neglect when multiple reports are received. SUMMARY OF PROVISIONS: Section 1.