Bill S6464 version original - Toby Ann Stavisky
BILL NUMBER: S6464 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the social services law, in relation to including phar- macists in the list of persons and officials required to report child abuse PURPOSE OR GENERAL IDEA OF BILL: To add pharmacists to the list of mandated reporters. SUMMARY OF PROVISIONS: Section 1. Amends section 413 of the Social Services law to add pharma- cists to the list of mandated reporter Section 2.