Bill S2943B version A - Brian Kavanagh
BILL NUMBER: S2943A SPONSOR: KAVANAGH TITLE OF BILL: An act to apply the Housing Stability and Tenant Protection Act of 2019 to rent calculations and rent records maintenance and destruction PURPOSE: This bill is intended to clearly define the prospective reach of certain provisions of the Housing Stability and Tenant Protection Act of 2019 in light of certain court decisions, including Regina Metro v. DHCR and Burrows v.