Bill S6489 version original - Timothy M. Kennedy
BILL NUMBER: S6489 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to licensure fees for seniors PURPOSE: This bill waives the fees for license renewal for any person who is sixty-two years of age or older. SUMMARY OF PROVISIONS: Section 1 amends paragraph c of subdivision 2 of section 503 of the vehicle and traffic law to waive the fee of license renewal for those who are sixty-two years of age or older.