Bill S6432A version A - Jack M. Martins
BILL NUMBER: S6432A SPONSOR: MARTINS TITLE OF BILL: An act in relation to authorizing the county of Nassau assessor to accept an application for a real property tax exemption from Torah Ohr Hebrew Academy PURPOSE: Authorizes the County of Nassau to grant a real property tax exemption to Torah Ohr Hebrew Academy. SUMMARY OF PROVISIONS: Section 1 of the bill authorizes the County of Nassau to accept an application for real property tax exemption for the property owned by Torah Ohr Hebrew Academy.