Bill S6491 version original - Brad Hoylman-Sigal
BILL NUMBER: S6491 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the civil practice law and rules, in relation to court confirmation of arbitration awards for public sector arbitrations SUMMARY OF PROVISIONS: Section 1. Technical amendment. Section 2. Provides that a court shall confirm an award upon the appli- cation of a party within one year after its delivery to the party unless an application to vacate or modify the award is made within 90 days after the delivery of the award.