Bill S6532 version original - Anthony H. Palumbo
BILL NUMBER: S6532 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the tax law, in relation to indexing taxes to inflation; and to repeal certain provisions of such law relating thereto PURPOSE: To index personal income taxes to the cost of living adjustment in order to provide taxpayers, families and businesses relief from record inflation. SUMMARY OF PROVISIONS: Section 1.