Bill S6541 version original - Robert Ortt
BILL NUMBER: S6541 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to creating the crimes of child abuse, vulnerable elderly person or incompetent or physically disabled person abuse, and torture PURPOSE OR GENERAL IDEA OF BILL: Enacts Noah's Law; creates the crimes of child abuse, vulnerable elderly person or incompetent or physically disabled person abuse, and torture. SUMMARY OF PROVISIONS: Section 1: This act shall be known and may be cited as "Noah's Law".