Bill S6558 version original - Anthony H. Palumbo
BILL NUMBER: S6558 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the environmental conservation law, in relation to the setback requirement in the Peconic and Gardiner's Bays for the Suffolk County Lease Program and Sugar Kelp PURPOSE: To change the setback requirement in the Peconic and Gardiners Bay for the Suffolk County Lease Program and Sugar Kelp production from 1,000 feet to 500 feet. SUMMARY OF PROVISIONS: Section 1.