Bill S6565 version original - Andrew J. Lanza
BILL NUMBER: S6565 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to making public lewdness in the first degree a class E felony PURPOSE OF BILL: This bill makes public lewdness in the first degree a Class E Felony. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the closing paragraph of section 245.03 of the penal law, as added by chapter 186 of the laws of 2014. Section two is the effective date.