Bill S6566 version original - Iwen Chu
BILL NUMBER: S6566 SPONSOR: CHU TITLE OF BILL: An act to amend the executive law, in relation to exempting certain business enterprise subcontractors who seek to do work with not-for-pro- fit corporations from certain requirements PURPOSE: The purpose of this bill is to provide an alternative method for the MWBE certification of small, ad-hoc, subcontractors of non-profits. SUMMARY OF PROVISIONS: Section 1.