Bill S6564 version original - Timothy M. Kennedy
BILL NUMBER: S6564 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the local finance law, in relation to facilitating the marketing of any issue of serial bonds or notes of the city of Buffalo issued on or before a certain date PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to extend the authorization for the private sale of bonds and notes of the city of from Buffalo June 30, 2023 to June 30, 2024. SUMMARY OF PROVISIONS: Section 1 amends section 54.