Bill S6552 version original - James Skoufis
BILL NUMBER: S6552 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public authorities law, in relation to establishing the Middletown parking authority and providing for its powers, duties and obligations; to repeal certain provisions of the public authorities law relating thereto; and providing for the repeal of such provisions upon expiration thereof PURPOSE: To provide for the creation of the City of Middletown Parking Authority. SUMMARY OF PROVISIONS: Section One.