Bill S6582 version original - Monica Martinez
BILL NUMBER: S6582 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the tax law, in relation to extending the authority of the county of Suffolk to impose an additional one percent of sales and compensating use tax PURPOSE: Extends Suffolk County's authority to impose their existing rate of sales and compensating use taxes for two years. SUMMARY OF PROVISIONS: Section 1.