Bill S6579A version original - Patrick M. Gallivan
BILL NUMBER: S6579 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as "Specialist Gerald B. Penn Memorial Bridge" PURPOSE OR GENERAL IDEA OF BILL: Designates the Portion of State route 219 crossing eighteen Mile Creek in the Town of Boston, county of Erie, as the "SP4 Gerald Penn. Memori- al Bridge.