Bill S6580 version original - Patrick M. Gallivan
BILL NUMBER: S6580 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Specialist Peter J. Loretto, Jr. Memo- rial Bridge" PURPOSE OR GENERAL IDEA OF BILL: Designates the bridge on State Route 75, crossing Eighteen Mile Creek, in the Town of Eden, county of Erie, as the "Specialist Peter J. Loret- to Jr. Memorial Bridge.