Bill S6572 version original - Rachel May
BILL NUMBER: S6572 SPONSOR: MAY TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization of the county of Cayuga to impose an additional one percent of sales and compensating use taxes PURPOSE: To authorize the County of Cayuga to impose a one percent sales tax SUMMARY OF PROVISIONS: Section 1 authorizes Cayuga County to impose a one percent sales tax through November 30, 2025. Section 2 is the effective date.