Bill S6578 version original - Patrick M. Gallivan
BILL NUMBER: S6578 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to desecration of property commemorating a historical figure PURPOSE: This legislation establishes the crime of desecration of property commemorating a historical figure to be a class C felony. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Adds a new section to the penal establishing the crime of desecration of property commemorating a historical figure to be a class C felony. Section 2.