Bill S6609 version original - Jeremy Cooney
BILL NUMBER: S6609 SPONSOR: COONEY TITLE OF BILL: An act to amend the executive law, in relation to the definition of small business for minority group members and women with respect to state contracts PURPOSE: Raises the number of workers required for consideration as an MWBE small business from 300 to 500 employees. SUMMARY OF SPECIFIC PROVISIONS: Section one of the bill amends the Executive Law to increase the thresh- old of workers from 300 to 500 for small business MWBEs.