Bill S6598 version original - Robert Jackson
BILL NUMBER: S6598 SPONSOR: JACKSON TITLE OF BILL: An act to amend the retirement and social security law, in relation to providing accidental disability retirement for deputy sheriffs in Suffolk county PURPOSE: This legislation clarifies situations in which a Deputy Sheriff in Suffolk County is entitled to accidental disability retirement and the amount of the disability retirement benefits. SUMMARY OF PROVISIONS: Section 1 adds a new Section 605-G to the Retirement and Social Security Law.