Bill S6220B version B - Toby Ann Stavisky
BILL NUMBER: S6220B SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to the practice of phys- ical therapy PURPOSE OR GENERAL IDEA OF BILL: To update the educational requirements for licensure as a physical ther- apist. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision b of section 6734 of the education law to require a doctoral degree in physical therapy for licensure.