Bill S6611 version original - Samra Brouk
BILL NUMBER: S6611 SPONSOR: BROUK TITLE OF BILL: An act to amend the executive law, in relation to including correctional facilities and local correctional facilities as publicly-assisted hous- ing accommodations for purposes of the division of human rights PURPOSE OR GENERAL IDEA OF BILL: This bill adds correctional facilities and law enforcement agencies to the jurisdiction of the human rights law. SUMMARY OF SPECIFIC PROVISIONS: Section one is the legislative intent and findings.