Bill S6622A version original - Michelle Hinchey
BILL NUMBER: S6622 SPONSOR: HINCHEY TITLE OF BILL: An act to allow Brian Laurange Jr. to be eligible to take the civil service exam for the position of police officer for the Columbia County Sheriff's Office PURPOSE: Authorizes Brian Laurange Jr. to take the competitive exam for full-time Police Officer and be placed on the eligible list for such position. SUMMARY OF PROVISIONS: Section 1. Authorizes Brian Laurange Jr.