Bill S6636 version original - Brad Hoylman-Sigal
BILL NUMBER: S6636 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the estates, powers and trusts law, in relation to the payment and distribution of damages in wrongful death actions PURPOSE OF BILL: This bill would permit the families of wrongful death victims to recover compensation for their emotional anguish. SUMMARY OF PROVISIONS OF BILL: Section 1 amends EPTL section 5-4.1 to extend the time permitted to bring a wrongful death action by one year and six months.