Bill S6633A version original - Anthony H. Palumbo
BILL NUMBER: S6633 SPONSOR: PALUMBO TITLE OF BILL: An act to authorize the county of Suffolk and the village of Nissequogue to offer an optional twenty-year retirement plan to police officer Bridget Topping PURPOSE: To authorize Suffolk County and Nissequogue Village to offer an optional twenty year retirement plan to Bridget Topping. SUMMARY OF PROVISIONS: Section 1.