Bill S6660 version original - Monica Martinez
BILL NUMBER: S6660 SPONSOR: MARTINEZ TITLE OF BILL: An act in relation to authorizing a study on lithium-ion battery fires and prevention PURPOSE: The purpose of this legislation is to authorize the department of fire prevention and control to conduct a study on lithium-ion battery fires and prevention. SUMMARY OF PROVISIONS: Section 1. authorizes the department of fire prevention and control to conduct a study on lithium-ion battery fires. Section 2.