Bill S6646 version original - Jessica Ramos
BILL NUMBER: S6646 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to requiring the commissioner of labor to prepare an annual report on the cost of living, poverty rates and adequacy of the current minimum wage in the state PURPOSE: To require the Commissioner of the Department of Labor to prepare an annual report on the state's cost of living, poverty rates, and the adequacy of the current minimum wage. SUMMARY OF PROVISIONS: Section 1.