Bill S6643A version original - Kevin S. Parker
BILL NUMBER: S6643 SPONSOR: PARKER TITLE OF BILL: An act to amend the correction law, in relation to increasing the amount of money furnished to incarcerated individuals upon release PURPOSE OR GENERAL IDEA OF THE BILL: This bill gives a bigger financial allowance to incarcerated individuals when they are released. SUMMARY OF PROVISIONS: Section 1: Subdivision 2 of section 125 of the correction law, as amended by chapter 322 of the laws of 2011, is amended.