Bill S6648 version original - James Sanders Jr.
BILL NUMBER: S6648 SPONSOR: SANDERS TITLE OF BILL: An act to amend the banking law, in relation to including participation in a banking development district as a factor when assessing a banking institution's record of performance PURPOSE: To accurately assess the performances. of a bank when in a development district. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the banking law to establish Subparagraph 10a. Section 2. Effective date.