Bill S1058A version A - Michelle Hinchey
BILL NUMBER: S1058A SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to estab- lishing the agrivoltaics viability pilot program PURPOSE OR GENERAL IDEA OF BILL: To construct, install, and operate dual-use solar energy projects on land which receive an agricultural assessment. SUMMARY OF PROVISIONS: Section 1: amends the agriculture and markets law by adding a new subsection to section 327 as legislative intent.