Bill S5475B version A - Robert Jackson
BILL NUMBER: S5475A SPONSOR: JACKSON TITLE OF BILL: An act to amend the labor law, in relation to inclusion of certain off- site custom fabrication as public work for the purposes of payment of prevailing wage PURPOSE: To ensure that construction and fabrication done off the public work site for specific use only in a public work project be compensated at the prevailing wage rate. SUMMARY OF PROVISIONS: Section 1 states the legislative findings and intent.