Bill S6671A version original - George Borrello
BILL NUMBER: S6671 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the tax law, in relation to authorizing the city of Olean to impose a hotel and motel tax PURPOSE: To authorize the city of Olean to impose a five percent hotel and motel tax. SUMMARY OF PROVISIONS: Section 1: the tax law is amended by adding section 1202-uu Section 2: Effective date.