Bill S6728 version original - Jessica Scarcella-Spanton
BILL NUMBER: S6728 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the judiciary law and the executive law, in relation to increasing the frequency of pretrial release data reporting PURPOSE OR GENERAL IDEA OF BILL: Bail is the amount of money defendants must post to be released from custody until their trial. Bail is not a fine. It is not supposed to be used as punishment.