Bill S6698A version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S6698 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act authorizing the Bais Tefila of Inwood to receive retroactive real property tax exempt status PURPOSE: To authorize the County of Nassau to grant a real property tax exemption to Bais Tefila of Inwood. SUMMARY OF PROVISIONS: Section 1: Authorizes the County of Nassau to accept a retroactive application for real property tax exemption for the property owned by Bais Tefila of Inwood in the town of Hempstead.