Bill S6700 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S6700 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act in relation to extending the boundaries of the Woodmere fire district to include the village of Woodsburgh PURPOSE: To allow the town of Hempstead to expand the Woodmere fire district's boundaries to include the village of Woodsburgh. SUMMARY OF PROVISIONS: Section I: Authorizes the town of Hempstead to expand the boundaries of the Woodmere fire district to cover and include the village of Woods- burgh.