Bill S6695 version original - Kevin S. Parker
BILL NUMBER: S6695 SPONSOR: PARKER TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of coercive control PURPOSE OR GENERAL IDEA OF BILL: Relates to establishing the crime of coercive control SUMMARY OF SPECIFIC PROVISIONS: Section 1: The penal law is amended by adding a new section 135.80. Section 2: This act shall take effect immediately.