Bill S6707 version original - William Weber
BILL NUMBER: S6707 SPONSOR: WEBER TITLE OF BILL: An act to amend the penal law, in relation to enacting "Paula's law" PURPOSE: "Paula's Law" is designed to ensure that persons who molest and then murder a child under the age of eighteen will never be released from prison. SUMMARY OF PROVISIONS: Section I: establishes the short title of this act, "Paula's law." Section II: states that in enacting this bill, it is the intent of the legislature to expand "Joan's Law.