Bill S6727 version original - Anthony H. Palumbo
BILL NUMBER: S6727 SPONSOR: PALUMBO TITLE OF BILL: An act in relation to the Southampton village ocean rescue PURPOSE: To authorize the village of Southampton to designate emergency response and rescue vehicles assigned to the Southampton Village Ocean Rescue as authorized emergency vehicles pursuant to section 101 of the vehicle and traffic law. SUMMARY OF PROVISIONS: Section 1.