Bill S6721 version original - Anthony H. Palumbo
BILL NUMBER: S6721 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the Indian law, in relation to the reinstatement of state recognition and acknowledgement of the Montaukett Indian Nation PURPOSE: Relates to the reinstatement of state recognition and acknowledgement of the Montaukett Indian Nation. It provides for the leadership of the Montaukett Indian Nation; qualification of voters and office SUMMARY OF SPECIFIC PROVISIONS: Section 1. Legislative Findings Section 2.