Bill S6737 version original - Steve Rhoads
BILL NUMBER: S6737 SPONSOR: RHOADS TITLE OF BILL: An act to amend the education law, in relation to restrictions on authority for the withholding of public monies from educational insti- tutions PURPOSE: Amends the education law, in relation to restrictions on authority for the withholding of public monies from educational institutions. SUMMARY OF PROVISIONS: Section 1.