Bill S6734 version original - Mark Walczyk
BILL NUMBER: S6734 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the public officers law, in relation to waiving the residency requirement for any appointed public official in Jefferson county PURPOSE: To waive the residency requirement for appointed public officials in the county of Jefferson. SUMMARY: Section 1 amends the public officials law to remove the residency requirement for public officials in the county of Jefferson, as so long they are a resident of an adjacent county.