Bill S6743A version original - Pete Harckham
BILL NUMBER: S6743 SPONSOR: HARCKHAM TITLE OF BILL: An act to authorize the town of Yorktown, county of Westchester, to discontinue as parklands and lease certain lands PURPOSE OR GENERAL IDEA OF BILL: Authorizes the town of Yorktown, county of Westchester, to discontinue as parklands and lease at fair market value certain lands. SUMMARY OF PROVISIONS: Section 1 authorizes Yorktown to lease land to Homeland Towers, LLC for the purposes of a radio transmission tower facility.