Bill S6731 version original - Roxanne J. Persaud
BILL NUMBER: S6731 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the penal law, in relation to the crime of computer tampering in the third degree PURPOSE: This legislation adds specific language to computer tampering in the third degree to cover instances when an individual enters, alters, or destroys computer materials related to the New York State and New York City immunization registry. SUMMARY OF PROVISIONS: Section one of the bill amends § 156.