Bill S1604F version B - Jessica Ramos
BILL NUMBER: S1604B SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to regulating the temperature of all indoor and outdoor worksites PURPOSE: This legislation regulates the temperature of indoor and outdoor work- sites within the scope of the article. SUMMARY OF PROVISIONS: Section 1. This act shall be known and may be cited as the "Worksite Temperature Regulation Act" Section 2. Legislative Findings and Intent Section 3. Amends labor law by adding Article 20-D 742.