Bill S5180A version A - Pete Harckham
BILL NUMBER: S5180A SPONSOR: HARCKHAM TITLE OF BILL: An act in relation to directing state agencies to adopt a sustainability and decarbonization program PURPOSE: Directs state agencies to adopt a sustainability and decarbonization program. SUMMARY OF PROVISIONS: Section 1 establishes definitions, and creates a GreenNY Council to design and implement training and outreach programs in order to met the requirements of this act.