Bill S6341C version A - Brad Hoylman-Sigal
BILL NUMBER: S6341A SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law, in relation to criminalizing the harass- ment of a child; and to amend the civil rights law, in relation to establishing a private right of action for harassment of a child PURPOSE: To protect children from harassment, stalking, and non-consensual recording. SUMMARY OF PROVISIONS: Section one adds to the penal law section 260.16 to create the class A misdemeanor of harassment of a child.