Bill S6110C version A - George Borrello
BILL NUMBER: S6110A SPONSOR: BORRELLO TITLE OF BILL: An act to amend the public officers law, in relation to waiving the local residency requirement for public officers in the village of Port- ville PURPOSE: Waives the local residency requirement for public officers in the village of Portville. SUMMARY OF PROVISIONS: Section 1.