Bill S6772 version original - Kevin S. Parker
BILL NUMBER: S6772 SPONSOR: PARKER TITLE OF BILL: An act to repeal subdivision 3 of section 3003 of the education law, relating to qualifications of superintendents PURPOSE OF BILL: This bill would eliminate the exception to the qualifications of super- intendents. SUMMARY OF SPECIFIC PROVISIONS: Section 3003 subdivision 3 of the Education Law would be repealed.