Bill S6780 version original - Kevin S. Parker
BILL NUMBER: S6780 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to the practice of licensed social workers PURPOSE: This bill will allow a person with a master's of social worker degree who meets certain requirements to file to be licensed within one year of the effective date of this act. SUMMARY OF PROVISIONS: Section 1.