Bill S6756 version original - John Liu
BILL NUMBER: S6756 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to clarifying what build- ings constitute an existing public school building for purposes of the location of charter schools SUMMARY OF PROVISIONS: Section 1 of the bill amends education law to clarify the definition of an existing public school building. Section 2 sets the effective date.