Bill S6751A version original - Timothy M. Kennedy
BILL NUMBER: S6751 SPONSOR: KENNEDY TITLE OF BILL: An act to authorize the widow of Gerald J. Sullivan to file an applica- tion with the New York state and local police and fire retirement system on behalf of such deceased member and deem such application as timely filed PURPOSE: This legislation would allow the widow of Gerald J. Sullivan to file an application with the New York state and local police and fire retirement system on behalf of Mr. Sullivan.