Bill S6763 version original - Joseph A. Griffo
BILL NUMBER: S6763 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the penal law, in relation to assaults on medical providers and employees and volunteers of hospitals PURPOSE: To include physicians, nurses, ambulance, directors, employees, agents and volunteers of hospitals performing an assigned duty within the Class D felony of assault in the second degree. SUMMARY OF PROVISIONS: 1. Amends subdivision 3 and 11 of section 120.