Bill S6784 version original - Kevin S. Parker
BILL NUMBER: S6784 SPONSOR: PARKER TITLE OF BILL: An act to amend the civil practice law and rules, in relation to creat- ing the anti-SLAPP act PURPOSE: Creates the anti-SLAPP act. SUMMARY OF PROVISIONS: Section 1: Entitles the act the "anti-SLAPP Act" Section 2: Amends subdivisions (c) and paragraph 1 of subdivision (g) of section 3211 of the civil practice law and rules.