Bill S2170A version A - John Liu
BILL NUMBER: S2170A SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to the participation of home-schooled pupils in certain school activities PURPOSE OR GENERAL IDEA OF BILL: To allow home-schooled students to participate in driver's education and athletic activities. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the education law to allow students to attend driver education courses in the school district where they reside.