Bill S6794 version original - Rachel May
BILL NUMBER: S6794 SPONSOR: MAY TITLE OF BILL: An act to amend the labor law, in relation to enacting the "trapped at work act" PURPOSE OR GENERAL IDEA OF BILL: To prohibit reimbursement clauses or promissory notes as a condition of employment. SUMMARY OF PROVISIONS: Section 1. Definitions. Section 2. (a) Rendering an employment promissory note unenforceable.