Bill S6790 version original - Anthony H. Palumbo
BILL NUMBER: S6790 SPONSOR: PALUMBO TITLE OF BILL: An act in relation to establishing a task force to create a comprehen- sive plan for the future of the Southampton campus of Stony Brook University PURPOSE OR GENERAL IDEA OF BILL: The intent of this legislation is to establish a task force to create a comprehensive plan for the future of the Southampton campus OF Stony Brook University. SUMMARY OF PROVISIONS: Section 1 describes the legislative findings and intent.