Bill S6808 version original - John Mannion
BILL NUMBER: S6808 SPONSOR: MANNION TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing first responder safety zones and speed limits in such safety zones PURPOSE: The purpose of this bill is to establish first responder safety zones in order to improve the safety of first responders entering the roadway in or near an emergency response station. SUMMARY OF PROVISIONS: Section 1 of this bill would establish first responder safety zones.