Bill S6810 version original - Cordell Cleare
BILL NUMBER: S6810 SPONSOR: CLEARE TITLE OF BILL: An act to amend the correction law, in relation to establishing a program to purchase fresh produce from farms in the state for incarcer- ated individuals PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to establish a program to purchase fresh produce from farms in the state for incarcerated individuals. SUMMARY OF PROVISIONS: Section 1.