Bill S6822 version original - Andrea Stewart-Cousins
BILL NUMBER: S6822 SPONSOR: STEWART-COUSINS TITLE OF BILL: An act to amend the tax law, in relation to extending the expiration of the mortgage recording tax imposed by the city of Yonkers PURPOSE: The purpose of this bill is to extend the expiration of the mortgage recording tax imposed by the City of Yonkers to August 31, 2025. SUMMARY OF PROVISIONS: Section 1 amends the tax law to extend the authority of the city of Yonkers to impose an additional mortgage recording tax until August 31, 2025.